Parser now in own method for use in channel parsing
